+++ lib/main.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
+import 'memoScreen.dart';
 
 void main() {
+// 最初に表示するWidget
   runApp(const MyApp());
 }
 
@@ -14,88 +16,10 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       title: 'My Todo App',
-      theme: ThemeData(
-        // テーマカラー
-        primarySwatch: Colors.blue,
-      ),
       // リスト一覧画面を表示
-      home: const memoScreen(),
-      // theme: ThemeData(
-      //   appBarTheme: const AppBarTheme(
-      //     color: Colors.white, // AppBarの背景色を白に設定
-      //     foregroundColor: Colors.black, // AppBarの文字色を黒に設定
-      //   ),
-      // ),
-      // home: const MyHomePage(title: 'メモ一覧'), // AppBarのタイトル
-    );
-  }
-}
-
-// メモ一覧画面用Widget
-class memoScreen extends StatelessWidget {
-  const memoScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        iconTheme: const IconThemeData(
-          color: Colors.black,
-          size: 30,
-        ),
-        title: const Text(
-          'メモ一覧',
-          textAlign: TextAlign.center,
-          style: TextStyle(
-            color: Colors.black,
-          ),
-        ),
-        centerTitle: true,
-        backgroundColor: Colors.white,
-      ),
-      body: const Center(
-        child: Text('リスト一覧画面'),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          // "push"で新規画面に遷移
-          Navigator.of(context).push(
-            MaterialPageRoute(builder: (context) {
-              // 遷移先の画面としてリスト追加画面を指定
-              return const TodoAddPage();
-            }),
-          );
-        },
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
-}
-
-// メモ追加画面用のWidget
-class TodoAddPage extends StatefulWidget {
-  const TodoAddPage({super.key});
-
-  @override
-  State<TodoAddPage> createState() => _TodoAddPageState();
-}
-
-class _TodoAddPageState extends State<TodoAddPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: TextButton(
-          // ボタンをクリックした時の処理
-          onPressed: () {
-            // "pop"で前の画面に戻る
-            Navigator.of(context).pop();
-          },
-          child: const Text('リスト追加画面（クリックで戻る）'),
-        ),
-      ),
+      home: memoScreen(),
     );
   }
 }
