+++ lib/memoAddition.dart
@@ -1,0 +1,38 @@
+// メモ追加画面用のWidget
+import 'package:flutter/material.dart';
+
+class TodoAddPage extends StatefulWidget {
+  const TodoAddPage({super.key});
+
+  @override
+  State<TodoAddPage> createState() => _TodoAddPageState();
+}
+
+class _TodoAddPageState extends State<TodoAddPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text(
+          'メモ追加',
+          textAlign: TextAlign.center,
+          style: TextStyle(
+            color: Colors.black,
+          ),
+        ),
+        centerTitle: true,
+        backgroundColor: Colors.white,
+      ),
+      body: Center(
+        child: TextButton(
+          // ボタンをクリックした時の処理
+          onPressed: () {
+            // "pop"で前の画面に戻る
+            Navigator.of(context).pop();
+          },
+          child: const Text('リスト追加画面（クリックで戻る）'),
+        ),
+      ),
+    );
+  }
+}
