+++ lib/memoScreen.dart
@@ -1,0 +1,45 @@
+// メモ一覧画面用のWidget
+import 'package:flutter/material.dart';
+import 'memoAddition.dart';
+
+class memoScreen extends StatelessWidget {
+  const memoScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text(
+          'メモ一覧',
+          textAlign: TextAlign.center,
+          style: TextStyle(
+            color: Colors.black,
+          ),
+        ),
+        centerTitle: true,
+        backgroundColor: Colors.white,
+      ),
+      body: ListView(
+        children: const [
+          Card(
+            child: ListTile(
+              title: Text(''),
+            ),
+          ),
+        ],
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          // "push"で新規画面に遷移
+          Navigator.of(context).push(
+            MaterialPageRoute(builder: (context) {
+              // 遷移先の画面としてリスト追加画面を指定
+              return const TodoAddPage();
+            }),
+          );
+        },
+        child: const Icon(Icons.add),
+      ),
+    );
+  }
+}
